requirements.txt
python-dotenv=0.13.0
requests=2.23.0
chart-studio==1.1.0
